backend/requirements.txt
# requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
pydantic
pydantic-settings

ccxt
redis

aio_pika

websockets

pytest
pytest-asyncio
aiosqlite
httpx

python-jose[cryptography]
passlib[argon2]
python-multipart